Final commit, everything done, might want to make everything compatible with GUI in the future. (delete, set) etc.

+++ src/main/java/me/imlukas/devnicschatplugin/channels/commands/ChannelCommand.java
@@ -1,6 +1,6 @@
-package me.imlukas.devnicschatplugin.commands;
+package me.imlukas.devnicschatplugin.channels.commands;
 
-import me.imlukas.devnicschatplugin.DevnicsChatPlugin;
+import me.imlukas.devnicschatplugin.ChannelsPlugin;
 import me.imlukas.devnicschatplugin.channels.Channel;
 import me.imlukas.devnicschatplugin.channels.DefaultChannels;
 import me.imlukas.devnicschatplugin.channels.config.ChannelConfig;
@@ -10,6 +10,8 @@ import me.imlukas.devnicschatplugin.sql.SQLHandler;
 import me.imlukas.devnicschatplugin.utils.TextUtil;
 import me.imlukas.devnicschatplugin.utils.storage.MessagesFile;
 import net.md_5.bungee.api.chat.ClickEvent;
+import net.md_5.bungee.api.chat.ComponentBuilder;
+import net.md_5.bungee.api.chat.HoverEvent;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.Bukkit;
 import org.bukkit.World;
@@ -20,20 +22,22 @@ import org.bukkit.command.TabCompleter;
 import org.bukkit.entity.Player;
 import org.bukkit.util.StringUtil;
 
-import java.rmi.server.UID;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
 
 
 public class ChannelCommand implements CommandExecutor, TabCompleter {
 
     private final static List<String> SUB_COMMANDS = List.of("delete", "create", "uuid", "reset", "set", "join");
-    private final DevnicsChatPlugin main;
+    private final ChannelsPlugin main;
     private final SQLHandler sqlHandler;
     private final ChannelConfig channelConfig;
     private final ChannelCache channelCache;
     private final MessagesFile messages;
 
-    public ChannelCommand(DevnicsChatPlugin main) {
+    public ChannelCommand(ChannelsPlugin main) {
         this.main = main;
         this.sqlHandler = main.getSqlHandler();
         this.channelConfig = main.getChannelConfig();
@@ -53,50 +57,55 @@ public class ChannelCommand implements CommandExecutor, TabCompleter {
             return true;
         }
 
-        if (!(player.hasPermission("channelmanager.admin"))) {
+        if (!(player.hasPermission("channelplugin.admin"))) {
             messages.sendMessage(player, "global.no-permission");
             return true;
         }
         if (args[0].equalsIgnoreCase("uuid")) {
             UUID channelUUID = channelCache.getChannel(player.getUniqueId());
 
-            TextComponent text = new TextComponent(TextUtil.color("&7Channel UUID: " + "&e" + channelUUID + "&7Click to copy"));
-            text.setClickEvent(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, channelUUID.toString()));
+            TextComponent text = new TextComponent(TextUtil.color("&7Channel UUID: " + "&e" + channelUUID));
+            TextComponent click = new TextComponent(TextUtil.color(" &7[Click to copy]"));
+            click.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to copy").create()));
+            click.setClickEvent(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, channelUUID.toString()));
 
-            player.spigot().sendMessage(text);
-            return true;
-        }
-        if (args[0].equalsIgnoreCase("create")) {
-            createChannel(player, args);
+            player.spigot().sendMessage(text, click);
             return true;
         }
 
-        if (args[0].equalsIgnoreCase("delete")) {
-            if (args.length == 2) {
+        if (args.length < 2) {
+            messages.sendMessage(player, "global.invalid-args");
+            return true;
+        }
+
+        switch (args[0]) {
+            case "create" -> {
+                createChannel(player, args);
+                return true;
+            }
+            case "delete" -> {
                 deleteChannel(player, args[1]);
                 return true;
             }
-            messages.sendStringMessage(player, "&c&l[Error]&7 Usage: /channel delete <channel-uuid>");
-            return true;
-        }
-        if (args[0].equalsIgnoreCase("join")) {
-            if (args.length == 2) {
+            case "join" -> {
                 joinChannel(player, args[1]);
                 return true;
             }
-            messages.sendStringMessage(player, "&c&l[Error]&7 Usage: /channel join <channel-uuid>");
-            return true;
+            case "reset" -> {
+                resetChannel(player);
+                return true;
+            }
         }
 
-        if (args[0].equalsIgnoreCase("reset")) {
-            channelCache.setPlayer(player.getUniqueId(), DefaultChannels.GLOBAL.channelUUID);
+        if (args.length < 3) {
+            messages.sendMessage(player, "global.invalid-args");
             return true;
         }
 
         Player target = Bukkit.getPlayer(args[1]);
 
         if (target == null) {
-            messages.sendMessage(player, "global.player-not-found");
+            messages.sendMessage(player, "global.player-not-found", (message) -> message.replace("%player%", args[1]));
             return true;
         }
 
@@ -111,6 +120,11 @@ public class ChannelCommand implements CommandExecutor, TabCompleter {
         return true;
     }
 
+    private void resetChannel(Player player){
+        channelCache.setPlayer(player.getUniqueId(), DefaultChannels.GLOBAL.channelUUID);
+        messages.sendMessage(player, "channel.reset");
+    }
+
     private void joinChannel(Player player, String UUID) {
         ChannelData channelData = getChannelData(player, UUID);
 
@@ -121,6 +135,7 @@ public class ChannelCommand implements CommandExecutor, TabCompleter {
 
         String channelName = channelData.getName();
         UUID channelUUID = channelData.getUUID();
+
         channelCache.setPlayer(player.getUniqueId(), channelUUID);
 
         messages.sendMessage(player, "channel.join", (message) -> message.replace("%channel%", channelName)
@@ -155,13 +170,15 @@ public class ChannelCommand implements CommandExecutor, TabCompleter {
         String channelName = channelData.getName();
         UUID channelUUID = channelData.getUUID();
 
-        channelConfig.deleteChannel(channelUUID);
-
-        if (channelConfig.getConfiguration().getConfigurationSection("channels." + channelUUID) != null) {
-            messages.sendMessage(player, "channel.delete-error", (message) -> message.replace("%channel%", channelName)
-                    .replace("%player%", player.getName()));
+        if (channelUUID.equals(DefaultChannels.GLOBAL.channelUUID)) {
+            messages.sendStringMessage(player, "&c&l[Error]&7 You can't delete the global channel!");
+            return;
         }
-        messages.sendMessage(player, "channel.delete-success", (message) -> message.replace("%channel%", channelName)
+        channelConfig.deleteChannel(channelUUID);
+        sqlHandler.resetPlayers(channelUUID);
+        channelCache.resetPlayers(channelUUID);
+
+        messages.sendMessage(player, "channel.delete", (message) -> message.replace("%channel%", channelName)
                 .replace("%player%", player.getName()));
 
     }
@@ -170,42 +187,41 @@ public class ChannelCommand implements CommandExecutor, TabCompleter {
     private void createChannel(Player player, String[] args) {
         Channel channel = new Channel(main);
 
+        int range = 0;
+        if (args.length > 3) {
+            range = parse(args[3]);
 
-        if (args.length == 2) {
-            channel.name(args[1])
-                    .create();
-            return;
+            if (range < 0) {
+                messages.sendStringMessage(player, "&c&l[Error]&7 Range must be bigger than 0 or equals to 0 (Unlimited range)");
+                return;
+            }
         }
-        if (args.length == 3) {
-            channel.name(args[1])
+
+        switch (args.length) {
+            case 2 -> channel.name(args[1])
+                    .prefix("&7[&e" + args[1] + "&7]&r")
+                    .create();
+
+            case 3 -> channel.name(args[1])
                     .prefix(args[2])
                     .create();
-            return;
-        }
 
-        int range = parse(args[3]);
-
-        if (range <= 0) {
-            messages.sendStringMessage(player, "&c&l[Error]&7 Range must be bigger than 0");
-            return;
-        }
-
-        if (args.length == 4) {
-            channel.name(args[1])
+            case 4 -> channel.name(args[1])
                     .prefix(args[2])
                     .range(range)
                     .create();
-        }
 
-        Set<World> worlds = new HashSet<>();
-        for (int i = 5; i < args.length; i++) {
-            worlds.add(Bukkit.getWorld(args[i]));
+            case 5 -> {
+                List<String> worlds = new ArrayList<>(Arrays.asList(args).subList(4, args.length));
+
+                channel.name(args[1])
+                        .prefix(args[2])
+                        .range(range)
+                        .worlds(worlds)
+                        .create();
+            }
         }
-        channel.name(args[1])
-                .prefix(args[2])
-                .range(range)
-                .worlds(worlds)
-                .create();
+        messages.sendMessage(player, "channel.create", (message) -> message.replace("%channel%", args[1]));
     }
 
     private ChannelData getChannelData(Player player, String UUID) {
@@ -239,42 +255,46 @@ public class ChannelCommand implements CommandExecutor, TabCompleter {
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String label, String[] args) {
         List<String> completions = new ArrayList<>();
-        if (args.length == 1) {
-            StringUtil.copyPartialMatches(args[0], SUB_COMMANDS, completions);
-        }
-        if (args.length == 2) {
-            if (args[0].equalsIgnoreCase("delete") || args[0].equalsIgnoreCase("join")) {
-                for (ChannelData channelData : channelConfig.getChannels().join()) {
-                    completions.add(channelData.getName());
+
+        switch (args.length) {
+
+            case 1 -> StringUtil.copyPartialMatches(args[0], SUB_COMMANDS, completions);
+            case 2 -> {
+                if (args[0].equalsIgnoreCase("delete") || args[0].equalsIgnoreCase("join")) {
+                    for (ChannelData channelData : channelConfig.getChannels().join()) {
+                        completions.add("" + channelData.getUUID() + " - " + channelData.getName());
+                    }
+                }
+                if (args[0].equalsIgnoreCase("set")) {
+                    for (Player player : Bukkit.getOnlinePlayers()) {
+                        completions.add(player.getName());
+                    }
+
+                }
+                if (args[0].equalsIgnoreCase("create")) {
+                    completions.add("<name>");
                 }
             }
-
-            if (args[0].equalsIgnoreCase("set")) {
-                for (Player player : Bukkit.getOnlinePlayers()) {
-                    completions.add(player.getName());
+            case 3 -> {
+                if (args[0].equalsIgnoreCase("set")) {
+                    for (ChannelData channelData : channelConfig.getChannels().join()) {
+                        completions.add("" + channelData.getUUID() + " - " + channelData.getName());
+                    }
+                }
+                if (args[0].equalsIgnoreCase("create")) {
+                    completions.add("<prefix>");
                 }
             }
-
-            if (args[0].equalsIgnoreCase("create")) {
-                completions.add("<name>");
+            case 4 -> {
+                if (args[0].equalsIgnoreCase("create")) {
+                    completions.add("<range>");
+                }
             }
-        }
-
-        if (args.length == 3) {
-            if (args[0].equalsIgnoreCase("create")) {
-                completions.add("<prefix>");
-            }
-        }
-        if (args.length == 4) {
-            if (args[0].equalsIgnoreCase("create")) {
-                completions.add("<range>");
-            }
-        }
-
-        if (args.length >= 5) {
-            if (args[0].equalsIgnoreCase("create")) {
-                for (World world : Bukkit.getWorlds()) {
-                    completions.add(world.getName());
+            case 5 -> {
+                if (args[0].equalsIgnoreCase("create")) {
+                    for (World world : Bukkit.getWorlds()) {
+                        completions.add(world.getName());
+                    }
                 }
             }
         }
